requirements.txt
fastapi~=0.85.0
Pillow~=9.2.0
uvicorn>=0.15.0,<0.16.0
img2vec_pytorch
python-multipart
